Refactor linter: extract pure logic to share between VSCode plugin and CLI
This refactoring creates a clean separation of concerns:

**Architecture:**
- BaseLinter: Pure linting logic, vscode-agnostic (used by both plugin + CLI)
  - lintSignal(), lintCommand(), lintCommands(), lintDocument()
  - No dependencies on vscode, UI frameworks, or any specific tool

- SignalLinter: VSCode plugin adapter
  - Extends BaseLinter (inherits pure linting methods)
  - Adds toDiagnostics() to convert LintResult → vscode.Diagnostic
  - Stores results for code actions

- CliSignalLinter: CLI adapter
  - Extends BaseLinter (inherits pure linting methods)
  - No framework-specific code needed

- vscodeAdapter.ts: VSCode-specific utilities (new)
  - mapToVscodeSeverity(): LintSeverity → vscode.DiagnosticSeverity
  - convertToVscodeDiagnostics(): LintResult[] → vscode.Diagnostic[]
  - Isolates all vscode imports to one file

**Benefits:**
- Single source of truth for all linting rules
- No duplication between plugin and CLI
- Easier to add new rules (implemented once, used everywhere)
- Clear separation: core logic vs. framework adapters
- Both VSCode plugin and CLI get the same behavior automatically
